Add a Tester application to test the model against existing data

+++ src/Modeler/Models.cs
@@ -1,6 +1,6 @@
 public class Issue
 {
-    public required uint Number { get; set; }
+    public required ulong Number { get; set; }
     public string? Label { get; set; }
     public required string Title { get; set; }
     public string? Body { get; set; }
